Take func_8005C62C's disc location as CdlLOC
func_8005C62C re-issues the movie stream's disc read. It declared its
parameter an integer and cast it back to a pointer three times, and both
callers took the address of a CdlLOC and cast that to s32 to make the call.

The parameter is now the location. The two call-site casts go, and the
prototype says what the function seeks to.

The three casts inside stay and are not the same thing: CdControlB takes its
parameter block as bytes, and DsRead2 takes DslLOC, which is the same four
fields under libds's name rather than libcd's. Those are library spellings,
not a missing type, and the header now says so.

The definition also spelled the parameter `int` while the header said `s32`;
only one of them now.

Byte-identical: make match, overlays and the full gate set pass.

+++ src/game/movie_frame_pipeline.c
@@ -223,7 +223,7 @@ s32 Movie_WaitAndDecodeFrame(s32 resync) {
             }
             timeouts++;
             if (resync != 0 || timeouts >= 10) {
-                func_8005C62C((s32)&D_8009B49C);
+                func_8005C62C(&D_8009B49C);
                 timeouts = 0;
             }
             /* One shared store: reorg copies it into both back-jump delay
